before:
```
name ............................... min time      avg time    std dv   runs
findAll ............................ 0.490 ms      0.925 ms    ±0.347  x1000
```
after:
```
name ............................... min time      avg time    std dv   runs
findAll ............................ 0.422 ms      0.782 ms    ±0.270  x1000
```